# search-and-recommendation/openapi.yaml
openapi: 3.0.3
info:
  title: Node Search, Location & Recommendation API
  version: 1.0.0
  description: Public API for Search, Location, and Recommendations (Node.js Service).
servers:
  - url: http://localhost:8086/api/v1
    description: Local development server
paths:
...
  /recommendations/home/feed:
    get:
      summary: Infinite scroll feed cho trang chủ
      parameters:
        - in: query
          name: lat
          schema:
            type: number
        - in: query
          name: lng
          schema:
            type: number
        - in: query
          name: category
          schema:
            type: string
            enum: [ALL, STAY, EXP, SVC]
            default: ALL
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Trang tiếp theo của feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
...
  /recommendations/provinces/{province}/destinations:
    get:
      summary: Discovery theo tỉnh
      description: Trả về Landmark và Complex/khu du lịch của tỉnh để người dùng đi tiếp vào detail hoặc listing quanh địa danh.
      parameters:
        - in: path
          name: province
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Landmark và Complex thuộc tỉnh
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Destination'
...
components:
  schemas:
...
    Destination:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [LANDMARK, COMPLEX]
        destinationType:
          type: string
          enum: [LANDMARK, COMPLEX]
        name:
          type: string
        province:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        thumbnailUrl:
          type: string
        radiusMeters:
          type: number
    FeedResponse:
      type: object
      properties:
        cursor:
          type: string
        nextCursor:
          type: string
        sections:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum:
                  [
                    LISTING_CAROUSEL,
                    DESTINATION_GRID,
                    HERO_LANDMARKS,
                    FEATURED_LANDMARKS,
                  ]
              title:
                type: string
              items:
                type: array
                items:
                  type: object
...